refactor(OrderTables): 주문 테이블 리스트를 일급 컬렉션으로 변경
